fix: use docker user as an image author and also tag latest
Previously, we were using the default image author, which is the IPFS author, which is initialized using the author of the service package, i.e., `elcollectooorr`. The release workflow was failing because we were instead trying to push the image using the docker user as an image author.

By introducing this change, we now use the docker user as the image author when building the image, and also tag and push the image with `latest`.

Please note that we might want to consider keeping the old approach when building the image, i.e., using the package author as the image author. In this case, we would push using this image name instead of the docker user. This might be the best approach as the image's name would contain the author of the package.

Moreover, if we were to keep the current commit's approach it might also make more sense to rename the `elcollectooorr` author of the packages.

[no ci]